Add files via upload
I conducted an exploratory data analysis project on the world's best-selling phones dataset from Kaggle. The objective was to clean, organize, and get insights pertaining to the sales of phones, manufacturers, and trends over time from this dataset. This project involved using SQL to handle inconsistencies in the data, transform it for analysis, and generate meaningful reports with insights.
This project gave an all-rounded analysis of best-selling phones in the world, highlighting several insights into the performance of manufacturers and product popularity by form factor, and trends over time. The insights derived from the analysis can, therefore, be used to inform business strategies, product development, and marketing efforts in the mobile phone industry.